feature: Add constant value MINIMUM_WINNING_COUNT with winning check method
